requirements.txt
altgraph==0.17.4
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
tk==0.1.0
